// apps/portal/src/api/generated/model/katastrImportResultResponseDto.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * HOA Manager API
 * The HOA Manager API description
 * OpenAPI spec version: 1.0
 */
import type { KatastrImportResultResponseDtoCounts } from "./katastrImportResultResponseDtoCounts";

export interface KatastrImportResultResponseDto {
    counts: KatastrImportResultResponseDtoCounts;
    effectiveAt: string;
}